requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
sympy>=1.13.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
